background-requirements.txt
Django
psycopg2>=2.7,<3.0
gunicorn==20.0.4
apprise
geopy
django-mailer
djangorestframework
